requirements.txt
setuptools
wheel
matplotlib
numpy
pandas
scipy
scikit-learn
psutil
seaborn
fuzzy-c-means
pyyaml
python-pptx
h5py
pyarrow
six
colorama
jinja2
joblib
hdbscan
pylint
black
pytz
pygod
pyod
tensorflow
keras
padasip
sphinx
furo
torch
pykalman
ortools
pulp
torch-geometric
torch-sparse
torch-scatter
torch-cluster
